// src_backup_pre_zustand/types.ts
export type RepeatMode = 'none' | 'once' | 'always' | 'folder';

export interface VideoItem {
  id: string;
  url: string;
  realPath?: string;
  title: string;
  repeatMode: RepeatMode;
  repeatCount: number;
  cols: number;
  folderFiles?: { name: string; url: string; path: string }[];
  currentIdx?: number;
  playing: boolean;
  muted: boolean;
  rotation?: number;
  currentTime?: number;
}

export interface TelemetryData {
  cpu: string;
  mem: string;
  gpu: string;
  temp?: number;
}